+++ VoidCrewTerminus/ReleaseFiles/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.12
+- lobby ship fix for non host playeres
+- fix forge module missing interaction targets
+- fix overlay missing on placed modules
 ## 0.0.11
 - update buildbox,relic,module's have hover ui to display mod info
 - update fab to display correct modded bulidbox recyle amount